utils: extract class 'Structure'
Extract the class 'Structure' from 'utils/win32/icon.py'
into 'utils/misc.py', as it provides a foundation for
extension. It is based on the struct module, which makes
it useful for different systems with different byte orders.
With this class any C-struct can be modelled.
